changed to record last 0.5 s of jitter time
There is some delay between the issuing of the IR camera record command
and the actual beginning of video recording. To ensure we don't miss
collecting any data, we decided to start the recording process 0.5 s
earlier. This means that we'll be recording the last 0.5 s of our
jitter interval, before our 1 s of background time before the trial
starts.

This behavior is under the control of a new parameter,
protocolParams.trialJitterRecordingDurationSec
-it adjusts the dataRecordingTime
-it makes the jitter interval shorter
-it makes the background window longer by the same amount